Fix attachment issue.
Previously, when a LaTeX body was too large, over 255 chars, when
Confluence attempted to save it to the database, it would throw
an exception because the table column for the Attachment Comment
was not large enough to hold the large LaTeX body. Added a length
check to avoid this by trimming the body if it is above 255 chars.

Also, upped version to 3.1