Fixes CRBSIT-455.
For pages that are not the latest version available, show an inlined
base64 version of the image rather than attaching the image to something.
Pretty sure that the attachments that were getting created were being
attached to the latest version of the page rather than the version
being accessed, so it would never actually show up on later visits
to the old page. This side-steps the problem by just inlining the
image.